Implement provider proc macro
